fix: Fix #22 Combining terms removes variable
The check to see if terms can be combined relied on a forEach
check on the varibales in the first term.
When the first term had no variables the checks were bypassed.

Added in a check for the length of the variables object for both terms to
ensure variables match.